refactor folder structure and a few variable names
