+++ freemocap/tests/test_mediapipe_skeleton_data_shape.py
@@ -3,7 +3,7 @@ import numpy as np
 from pathlib import Path
 from typing import Union
 
-from freemocap.tests.utilities.get_number_of_frames_of_videos_in_a_folder import (
+from freemocap.utilities.get_number_of_frames_of_videos_in_a_folder import (
     get_number_of_frames_of_videos_in_a_folder,
 )
