+++ freemocap/utilities/download_sample_data.py
@@ -1,9 +1,9 @@
 import io
 import logging
-
-import requests
 import zipfile
 from pathlib import Path
+
+import requests
 
 from freemocap.system.paths_and_filenames.file_and_folder_names import (
     FREEMOCAP_SAMPLE_DATA_RECORDING_NAME,
@@ -13,10 +13,19 @@ from freemocap.system.paths_and_filenames.path_getters import get_recording_sess
 
 logger = logging.getLogger(__name__)
 
-# TODO - some of the naming in this file is inconsistent with the rest of the codebase. Needs fixed at some point. Also I think a lot of the `find_..` functions should be in a different file?
+def get_sample_data_path(download_if_needed:bool=True)->str:
+
+    sample_data_path =  str(Path(get_recording_session_folder_path()) / FREEMOCAP_SAMPLE_DATA_RECORDING_NAME)
+    if not Path(sample_data_path).exists():
+        if download_if_needed:
+            download_sample_data()
+        else:
+            raise Exception(f"Could not find sample data at {sample_data_path} (and `download_if_needed` is False)")
+
+    return sample_data_path
 
 
-def load_sample_data(sample_data_zip_file_url: str = FIGSHARE_ZIP_FILE_URL) -> str:
+def download_sample_data(sample_data_zip_file_url: str = FIGSHARE_ZIP_FILE_URL) -> str:
     try:
         logger.info(f"Downloading sample data from {sample_data_zip_file_url}...")
 
@@ -40,5 +49,5 @@ def load_sample_data(sample_data_zip_file_url: str = FIGSHARE_ZIP_FILE_URL) -> s
 
 
 if __name__ == "__main__":
-    sample_data_path = load_sample_data()
+    sample_data_path = download_sample_data()
     print(f"Sample data downloaded successfully to path: {str(sample_data_path)}")
